+++ src/types.h
@@ -14,7 +14,7 @@ typedef uint32_t uint32;
 typedef uint64_t uint64;
 
 typedef float real32;
-typedef float real64;
+typedef double real64;
 
 typedef uint32_t bool32;
